Refactor: register 컴포넌트 쪼개기 리팩토링
