+++ app/src/components/ImagaeUploadForm.jsx
@@ -1,0 +1,39 @@
+import { useCallback, useEffect, useRef, useState } from "react";
+import ImageItem from "./ImageItem";
+import ImageList from "./ImageList";
+import RegisterFormCategory from "./RegisterFormCategory";
+import RegisterFormTitle from "./RegisterFormTitle";
+import RegisterFormUpload from "./RegisterFormUpload";
+const ImageUploadForm = ({ maxImageNum, setCategory, title, setTitle, items, setItems }) => {
+  const refCategory = useRef();
+
+  const selectHandler = useCallback((e) => {
+    setCategory(e.target.value);
+  }, []);
+
+  const onImageItemSetting = (imageTitle, tags, i) => {
+    setItems(items.map((item) => (item.id === i ? { ...item, title: imageTitle, tags } : item)));
+    console.log(items);
+  };
+
+  const deleteItem = (index) => {
+    setItems(items.filter((item, i) => index !== i));
+  };
+
+  useEffect(() => {
+    for (let j = 0; j < items.length; j++) {
+      items[j].id = j;
+    }
+  }, [items]);
+
+  return (
+    <main>
+      <RegisterFormTitle title={title} setTitle={setTitle} refCategory={refCategory} />
+      <RegisterFormCategory selectHandler={selectHandler} refCategory={refCategory} />
+      <RegisterFormUpload maxImageNum={maxImageNum} items={items} setItems={setItems} />
+      <ImageList items={items} onImageItemSetting={onImageItemSetting} deleteItem={deleteItem} />
+    </main>
+  );
+};
+
+export default ImageUploadForm;
